+++ app/src/components/CardsComponent.vue
@@ -1,5 +1,7 @@
 <script lang="ts">
-  interface CardItem {
+  import CardItem from './CardItem.vue';
+
+  interface Cards {
     name: string,
     contents: Array<string>,
     featureList: Array<Feature>,
@@ -12,39 +14,27 @@
 
   export default {
     props: {
-      cardItems: Array<CardItem>,
+      cardItems: Array<Cards>,
+    },
+    components: {
+      CardItem,
     }
   }
 </script>
 
 <template>
   <div class="item-list-wrapper">
-    <div
+    <CardItem
       v-for="cardItem in cardItems"
       :key="cardItem.name"
-      class="item"
-    >
-      <h4>{{ cardItem.name }}</h4>
-      <p v-for="content in cardItem.contents" :key="content">{{ content }}</p>
-      <ul v-if="cardItem.featureList">
-        <li v-for="feature in cardItem.featureList" :key="feature.heading">
-          <strong>{{ feature.heading }}</strong>
-          <span v-if="feature.content"> - {{ feature.content }}</span>
-        </li>
-      </ul>
-    </div>
+      :cardItem="cardItem"
+    />
   </div>
 </template>
 
 <style>
   .item-list-wrapper {
     margin-top: 1rem;
-  }
-
-  .item {
-    border: 1px solid var(--color-primary-dark);
-    padding: 1rem;
-    margin-bottom: 1rem;
   }
 
   @media(min-width: 992px) {